requirements.txt
python-whois
dnspython
requests[socks]
idna
pysocks
